Added a settings menu and map button
The settings menu is a separate activity called SettingsActivity that has
options for postal code location and temperature units. Also, a button
was added to the action bar in the MainActivity that lets the user view
their chosen postal code location in a maps app. Note that the
SettingsActivity is implemented in a way that is now deprecated in order
to maintain compatibility with Gingerbread.

- Added activity for settings menu with options for location and
temperature units
- Added map button