Added a share button to the detailed view's action bar
The share button is a standard share button utilizing the ShareActionProvider.
This creates a share button that populates a list of all apps that can share the
data passed in the share intent. It also keeps a button for quick access to the last
app the user selected to share.

- Added share button to detailed view