Added tests for database entries
Files in app/src/androidTest contain the tests. They test
insertions and queries to the weather forecasts table and
to the location table. Test by right-clicking one of the
directories containing tests and running from there, or by
running the Full Test Suite from the Run button in the
toolbar.

- Added tests for inserting and querying from weather table
and from location table